+++ deepcell_types/__init__.py
@@ -3,3 +3,10 @@ from .dct_kit.config import DCTConfig as DCTConfig
 from .predict import predict as predict
 from .predict import PredictionResult as PredictionResult
 from .preprocessing import preprocess_fov as preprocess_fov
+
+__all__ = [
+    "predict",
+    "PredictionResult",
+    "DCTConfig",
+    "preprocess_fov",
+]