+++ docs/conf.py
@@ -9,7 +9,7 @@
 project = 'deepcell-types'
 copyright = '%Y, Van Valen Lab at the California Institute of Technology (Caltech)'
 author = 'Xuefei (Julie) Wang'
-release = '0.0.1-dev'
+release = '0.1.0'
 
 # -- General configuration ---------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration